Fixed text-decoration-thickness rendering
CSS text-decoration-thickness renders 1px thinner compared to other
browsers (Firefox, Safari),
when the calculated thickness has a decimal part >= .5px,
because the value gets floored by
`GraphicsContext::DrawLineForText` via `RoundDownThickness`.

The other engines seem to round to the nearest integer when
text-decoration-thickness value is not auto.

The proposed fix is to match the other engines by rounding the
text-decoration-thickness value when it is not auto.
The rounding is done in `ComputeDecorationThickness` to avoid also
rounding when text-decoration-thickness value is set to auto.

Tentative WPT tests are added to reflect the rounding behavior.

R=dbaron, futhark, pdr, schenney

Change-Id: Ie26664884d2cc48d2b7ffd838a21d0b771303718
Bug: 1255280