Refactor tests to remove dependency on offsets.
Remove start, end, and scroll-offset from scroll-linked animation
CSS tests.  The parameters "start" and "end" do not align with the
current spec, and "scroll-offset" is likely to be phased out in the
spec rewrite.

A consistent scroll range can be obtained by using overflow: hidden
and a ratio of the content height to container height.

Bug: 1326242
Change-Id: I0925ae0dbf323cc409dd7f3a272b22a6d9e17525